+++ aql-engine/src/test/java/org/ehrbase/openehr/aqlengine/sql/AqlSqlQueryBuilderTest.java
@@ -25,6 +25,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import org.ehrbase.api.knowledge.KnowledgeCacheService;
+import org.ehrbase.api.service.TemplateService;
 import org.ehrbase.openehr.aqlengine.asl.AqlSqlLayer;
 import org.ehrbase.openehr.aqlengine.asl.AslGraph;
 import org.ehrbase.openehr.aqlengine.asl.model.query.AslRootQuery;
@@ -52,10 +53,14 @@ class AqlSqlQueryBuilderTest {
     private final KnowledgeCacheService mockKnowledgeCacheService = mock();
 
     private AqlSqlQueryBuilder aqlSqlQueryBuilder() {
+        TemplateService templateService = mock();
+        Mockito.when(templateService.findAllTemplateIds())
+                .thenReturn(Map.of(UUID.randomUUID(), "template1.v1", UUID.randomUUID(), "template2.v3"));
+
         return new AqlSqlQueryBuilder(
                 TestConfig.aqlConfigurationProperties(),
                 new DefaultDSLContext(SQLDialect.POSTGRES),
-                mockKnowledgeCacheService,
+                templateService,
                 Optional.empty());
     }
 
@@ -78,7 +83,8 @@ class AqlSqlQueryBuilderTest {
               CONTAINS FOLDER[name/value="Encounter"]
               CONTAINS FOLDER enc
               CONTAINS COMPOSITION c
-            WHERE e/ehr_id/value = 'e6fad8ba-fb4f-46a2-bf82-66edb43f142f'
+            WHERE e/ehr_id/value = 'e6fad8ba-fb4f-46a2-bf82-66edb43f142f' and c/archetype_details/template_id/value = 'template1.v1'
+            order by c/archetype_details/template_id/value
         """);
 
         System.out.println("/*");
@@ -86,9 +92,6 @@ class AqlSqlQueryBuilderTest {
         System.out.println("*/");
 
         AqlQueryWrapper queryWrapper = AqlQueryWrapper.create(aqlQuery);
-
-        Mockito.when(mockKnowledgeCacheService.findUuidByTemplateId(ArgumentMatchers.anyString()))
-                .thenReturn(Optional.of(UUID.randomUUID()));
 
         AqlSqlLayer aqlSqlLayer = new AqlSqlLayer(mockKnowledgeCacheService, () -> "node");
         AslRootQuery aslQuery = aqlSqlLayer.buildAslRootQuery(queryWrapper);
@@ -117,8 +120,6 @@ class AqlSqlQueryBuilderTest {
 
         AqlQuery aqlQuery = AqlQueryParser.parse(aql);
         AqlQueryWrapper queryWrapper = AqlQueryWrapper.create(aqlQuery);
-        Mockito.when(mockKnowledgeCacheService.findUuidByTemplateId(ArgumentMatchers.anyString()))
-                .thenReturn(Optional.of(UUID.randomUUID()));
 
         AqlSqlLayer aqlSqlLayer = new AqlSqlLayer(mockKnowledgeCacheService, () -> "node");
         AslRootQuery aslQuery = aqlSqlLayer.buildAslRootQuery(queryWrapper);