+++ service/src/main/java/org/ehrbase/service/CompositionServiceImp.java
@@ -395,7 +395,7 @@ public class CompositionServiceImp implements CompositionService {
                 if (templateId == null) {
                     return Optional.empty();
                 }
-                return Optional.ofNullable(knowledgeCacheService.getQueryOptMetaData(templateId));
+                return Optional.ofNullable(knowledgeCacheService.getWebTemplate(templateId));
             }
         };
     }